feat(lead-engineer): extract PlanProcessor + add knowledge-augmented plan search
Extract PlanProcessor from lead-engineer-processors.ts into dedicated
lead-engineer-plan-processor.ts with knowledge search augmentation:

- libs/prompts/src/lead-engineer-prompts.ts: new buildPlanPrompt() that
  accepts optional priorPlans context and injects 'Prior Plans for Similar
  Features' section into the prompt
- libs/prompts/src/index.ts: export buildPlanPrompt
- apps/server/src/services/lead-engineer-plan-processor.ts: extracted
  PlanProcessor with knowledgeStoreService.search() call before plan gen;
  non-blocking — proceeds normally if search returns empty or throws
- apps/server/src/services/lead-engineer-processors.ts: remove PlanProcessor
  (now in dedicated file); retains IntakeProcessor only
- apps/server/src/services/lead-engineer-state-machine.ts: import
  PlanProcessor from new dedicated module
