+++ apps/server/src/services/lead-engineer-state-machine.ts
@@ -9,7 +9,8 @@ import { createLogger } from '@protolabs-ai/utils';
 import type { Feature, ExecuteOptions, GoalGateResult, EventType } from '@protolabs-ai/types';
 import type { EventEmitter } from '../lib/events.js';
 import type { PipelineCheckpointService } from './pipeline-checkpoint-service.js';
-import { IntakeProcessor, PlanProcessor } from './lead-engineer-processors.js';
+import { IntakeProcessor } from './lead-engineer-processors.js';
+import { PlanProcessor } from './lead-engineer-plan-processor.js';
 import { ExecuteProcessor } from './lead-engineer-execute-processor.js';
 import { ReviewProcessor, MergeProcessor } from './lead-engineer-review-merge-processors.js';
 import { DeployProcessor } from './lead-engineer-deploy-processor.js';
